set up sockets in a re-usable way
