+++ islands/serviceMap.tsx
@@ -17,18 +17,13 @@ import { Audience } from "streamdal-protos/protos/sp_common.ts";
 import { Pipeline } from "streamdal-protos/protos/sp_pipeline.ts";
 import { FlowEdge, FlowNode } from "../lib/nodeMapper.ts";
 import { serviceSignal } from "../components/serviceMap/serviceSignal.ts";
-import { useEffect, useRef, useState } from "preact/hooks";
+import { useRef, useState } from "preact/hooks";
 import { OP_MODAL_WIDTH } from "./drawer/infoDrawer.tsx";
 import { EmptyService } from "../components/serviceMap/emptyService.tsx";
 import {
   ComponentEdge,
   ServiceEdge,
 } from "../components/serviceMap/customEdge.tsx";
-import {
-  audienceMetricsSocket,
-  serverErrorSocket,
-  serviceMapSocket,
-} from "../lib/sockets.ts";
 import {
   OP_MODAL_KEY,
   opModal,
@@ -113,17 +108,6 @@ export default function ServiceMapComponent(
   const wrapper = useRef(null);
 
   const [rfInstance, setRfInstance] = useState(null);
-
-  useEffect(() => {
-    const serviceSocket = serviceMapSocket("./ws/service-map");
-    const audienceSocket = audienceMetricsSocket("./ws/audience-metrics");
-    const errorSocket = serverErrorSocket("./ws/server-error");
-    return () => {
-      serviceSocket?.close();
-      audienceSocket?.close();
-      errorSocket?.close();
-    };
-  }, []);
 
   const [nodes, setNodes, onNodesChange] = useNodesState(initNodes);
   const [edges, setEdges, onEdgesChange] = useEdgesState(initEdges);
